# openapi/specs/companies.yaml
openapi: 3.0.0
info:
  title: VICTA API - Companies
  version: "1.0.0"
servers:
  - url: /api/v1
paths:
  /companies:
    post:
      summary: Создание компании
      operationId: createCompany
      security:
        - bearerAuth: [ ]
      responses:
        "200":
          description: Компания создана
          content:
            application/json:
              schema:
                $ref: "../components/api_response.yaml"
    get:
      summary: Получение списка компаний
      operationId: getCompanies
      security:
        - bearerAuth: [ ]
      responses:
        "200":
          description: Список компаний
          content:
            application/json:
              schema:
                $ref: "../components/api_response.yaml"
  /companies/{id}:
    get:
      summary: Получение компании по id
      operationId: getCompany
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Данные компании
          content:
            application/json:
              schema:
                $ref: "../components/api_response.yaml"
    put:
      summary: Обновление компании
      operationId: updateCompany
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Компания обновлена
          content:
            application/json:
              schema:
                $ref: "../components/api_response.yaml"
    delete:
      summary: Удаление компании
      operationId: deleteCompany
      security:
        - bearerAuth: [ ]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Компания удалена
          content:
            application/json:
              schema:
                $ref: "../components/api_response.yaml"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer